Add extra condition for creating suggestions on input of 1 word
Normally we'll not create suggestions on a 1 word only input, but in the
loop, we'll check if there's actually more input to come (in the same
context), if there is, we'll generate suggestions on that 1 word input.
This is to handle the case when the error occurs on the first word.